backend/requirements.txt
fastapi
uvicorn
numpy
pydantic
httpx
python-jose[cryptography]
python-dotenv
google-genai>=1.0.0
resend

# Testing
pytest
